rename file and fix styling
